Create entrance to the maze
The entrance is one of the furthest cells in the maze which is adjacent
to a cell not in the maze. The maze drawing tool shows the entrance in
red. Note that the entrance may require that the player goes above or
below the maze to reach, but as long as the player is able to get on top
the entrance will always be accessible.